% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRviaSPCN-package.R
\docType{package}
\name{DRviaSPCN-package}
\alias{DRviaSPCN-package}
\alias{DRviaSPCN}
\title{The DRviaSPCN package}
\description{
DRviaSPCN is a R package for repurposing drugs via subpathway crosstalk network.
}
\details{
DRviaSPCN

The main goals of the DRviaSPCN package is to provide main functions and some data.

         The main functions includes: 1) DE2SubPath 2) getSubpathscore 3) optimaldrugs and four visualization functions: 1) plotSPW 2) getMolecularFM 3) Disease2SPWheatmap 4) Drug2SPWheatmap.

         For more details, please see browseVignettes("DRviaSPCN")
}
