% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{envData}
\alias{envData}
\title{An environment variable which includes some example data}
\format{
An environment variable
}
\usage{
envData
}
\description{
An environment variable which includes some example data.
DE2SubPathresult: The result of function "DE2SubPath" with "perm = FALSE".
DE2SubPathresult_P: The result of function "DE2SubPath" with "perm = TRUE".
Drugs_CID: PubCham Database ID of drugs.
GEP: An example gene expression profile.
GOInfo: Biological Process data from Gene Ontology.
GoSubPconGene: Genes symble shared by a subpathway with Biological Process.
heatmap.list: The result of function "Drug2SPWheatmap".
Jaccardscore: Jaccard score shared by a subpathway with Biological Process.
label: A character vector consist of "0" and "1" which represent sample class in gene expression profile.
Opdrugresult: The result of function "optimaldrugs".
SubPathwayInfo: Subpathway information from SubpathwayMiner.
SubPathwaymapdata: Subpathway structure data.
}
\keyword{datasets}
