% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEscore.R
\name{getDEscore}
\alias{getDEscore}
\title{Calculating Log2 Fold Change of genes}
\usage{
getDEscore(inexpData, Label)
}
\arguments{
\item{inexpData}{A gene expression profile of interest (rows are genes, columns are samples).The data in the expression profile is best not be log2 converted.}

\item{Label}{A character vector consist of "0" and "1" which represent sample class in gene expression profile. "0" means normal sample and "1" means disease sample.}
}
\value{
A one-column matrix of Log2 Fold Change which rownames is gene.
}
\description{
Function "getDEscore" uses gene expression profile to calculate Log2 Fold Change of genes.
}
