% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExample.R
\name{GetExample}
\alias{GetExample}
\title{Get example data}
\usage{
GetExample(exampleData)
}
\arguments{
\item{exampleData}{A character, should be one of"GEP","Slabel",
"CentralityScoreResult","Opdrugresult" and "heatmap.list".}
}
\value{
example data
}
\description{
This function is used to achieve exxample data.
}
