% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageLoaded.R
\name{PackageLoaded}
\alias{PackageLoaded}
\title{PackageLoaded}
\usage{
PackageLoaded(name)
}
\arguments{
\item{name}{A character which is the name of package.}
}
\value{
A boolean value.
}
\description{
Determine if the package is loaded, if no package is loaded.
}
