% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{envData}
\alias{envData}
\title{An environment variable which includes some example data}
\format{
An environment variable
}
\usage{
envData
}
\description{
An environment variable which includes some example data.
CentralityScoreResult: The result of function "CalCentralityScore".
Drugs_CID: PubCham Database ID of drugs.
GEP: An example gene expression profile.
GOInfo: Biological Process data from Gene Ontology.
GoSubPconGene: Genes symble shared by a pair of subpathways with Biological Process.
Jaccardscore: Jaccard score shared by a pair of subpathways with Biological Process.
Slabel: A character vector consist of "0" and "1" which represent sample class in gene expression profile.
Opdrugresult: The result of function "Optimaldrugs".
SubPathwayInfo: Subpathway information from SubpathwayMiner.
SubPathwaymapdata: Subpathway structure data.
}
\keyword{datasets}
