% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMolecularFm.R
\name{getMolecularFm}
\alias{getMolecularFm}
\title{Plot chemical molecular formula  of drugs}
\usage{
getMolecularFm(drugname = "", main = "", sub = "")
}
\arguments{
\item{drugname}{A character string of drug name.}

\item{main}{An overall title for the chemical structure graph.}

\item{sub}{A sub title for the chemical structure graph.}
}
\value{
Chemical molecular formula  of the drug or compound.
}
\description{
The function "getMolecularFm" outputs the chemical molecular formula  of a drug or compound . The results can be visualized by the "plot" function.
}
\examples{
##Load depend package
library(ChemmineR)
library(rvest)
# Obtain molecular formula and visualize it.
Mole_formula<-getMolecularFm(drugname ="methotrexate")
plot(Mole_formula)
}
